🚚 Move: ProfileLink 를 shared → entities/profile 로
FSD 레이어 위반이었다. `shared/ui/ProfileLink` 가 `entities/profile/model/displayName`
의 `profileHref` 를 import 하고 있었다 — shared 는 상위 레이어를 참조할 수 없다.

`profileHref` 를 shared 로 내리는 방향도 있었지만 그러지 않았다. 이 함수가 담은 건
"공개 서재 경로는 /[username] 이고 조회도 username 컬럼으로만 된다" 는 프로필 도메인
지식이라, shared 로 내리면 그 지식이 레이어 밖으로 샌다. 그래서 컴포넌트를 올렸다.

entities 에 `ui/` 가 생기는 첫 사례다(기존 entities 는 api·model·types 만 있었다).
FSD 표준 슬라이스이고, 이 컴포넌트는 도메인 지식을 쓰는 UI 라 여기가 제자리다.

사용처 4곳(CommentItem·UserCard·ActivityCard·ReaderProfiles)의 import 경로를 바꿨다.
검증: lint 0 · tsc 0 · 테스트 280 통과.

발견 경위: Storybook 도입 중 `src/shared/ui/**` 가 lint globalIgnores 라
자체 컴포넌트까지 검사에서 빠져 있는 걸 확인했고, 무시를 풀자 이 위반이 드러났다.
